# Improve performance of bundleobj on macOS