--- a/framework/test/gtest/BundleObjFileTest.cpp
+++ b/framework/test/gtest/BundleObjFileTest.cpp
@@ -25,6 +25,8 @@
 
 #include "cppmicroservices/util/FileSystem.h"
 
+#include "cppmicroservices/util/MappedFile.h"
+
 #include "TestUtils.h"
 #include "TestingConfig.h"
 
@@ -76,3 +78,28 @@ TEST(BundleObjFile, GetRawBundleResourceContainer)
 #endif
 }
 
+#if defined (US_BUILD_SHARED_LIBS)
+#if defined (US_PLATFORM_APPLE) || defined (US_PLATFORM_POSIX)
+TEST(BundleObjFile, MappedFile)
+{
+  int fileDesc = open(testBundlePath.c_str(), O_RDONLY);
+  struct stat sb;
+  fstat(fileDesc, &sb);
+  off_t offset{0};
+  off_t pa_offset = offset & ~(sysconf(_SC_PAGE_SIZE) - 1);
+  /* offset for mmap() must be page aligned */
+  size_t length = sb.st_size - offset;
+  close(fileDesc);
+
+  cppmicroservices::MappedFile mappedBundleFile(testBundlePath, length, pa_offset);
+  ASSERT_TRUE(mappedBundleFile.GetData());
+  ASSERT_GT(mappedBundleFile.GetSize(), 0u);
+
+  ASSERT_NO_THROW({
+      cppmicroservices::MappedFile mappedBundleFile("/does/not/exist/bogus.bundle", 0, 0);
+      ASSERT_EQ(mappedBundleFile.GetData(), nullptr);
+      ASSERT_EQ(mappedBundleFile.GetSize(), 0u);
+  });
+}
+#endif // defined (US_PLATFORM_APPLE) || defined (US_PLATFORM_POSIX)
+#endif // defined (US_BUILD_SHARED_LIBS)